wizard/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
jinja2>=3.1
python-multipart>=0.0.9
pyyaml>=6.0
httpx>=0.27
aiofiles>=23.2
defusedxml>=0.7
